Standardize solver API and implement cross-platform loadBook architecture
- Standardized BaseConnect4Solver API by unifying async analysis methods, properties (width/height), and book loading across all Sync/Async/Native/Threaded implementations.
- Implemented robust loadBook functionality directly into Android (JNI/Kotlin) and iOS (Obj-C/Swift) bindings, allowing native consumers to fully utilize the C++ cached opening books without WASM translation overhead.
- Added extensive caching integration tests to verify accurate interception of mocked invalid book scores.

+++ native/node_binding.cpp
@@ -1530,7 +1530,7 @@ Value GetNodeCount(const CallbackInfo& info) {
         else if (w == 13 && h == 8) count = static_cast<C4_13x8::Solver*>(solver)->getNodeCount();
         else if (w == 9 && h == 9) count = static_cast<C4_9x9::Solver*>(solver)->getNodeCount();
     }
-    return Number::New(info.Env(), (double)(uintptr_t)count);
+    return Number::New(info.Env(), (double)count);
 }
 
 // Helper macro for book operations
